+++ src/main/java/io/supertokens/storage/mysql/queries/GeneralQueries.java
@@ -26,6 +26,7 @@ import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.pluginInterface.multitenancy.TenantIdentifier;
 import io.supertokens.storage.mysql.ConnectionPool;
+import io.supertokens.storage.mysql.PreparedStatementValueSetter;
 import io.supertokens.storage.mysql.Start;
 import io.supertokens.storage.mysql.config.Config;
 import io.supertokens.storage.mysql.utils.Utils;
@@ -43,8 +44,7 @@ import java.util.stream.Collectors;
 import static io.supertokens.storage.mysql.PreparedStatementValueSetter.NO_OP_SETTER;
 import static io.supertokens.storage.mysql.ProcessState.PROCESS_STATE.CREATING_NEW_TABLE;
 import static io.supertokens.storage.mysql.ProcessState.getInstance;
-import static io.supertokens.storage.mysql.QueryExecutorTemplate.execute;
-import static io.supertokens.storage.mysql.QueryExecutorTemplate.update;
+import static io.supertokens.storage.mysql.QueryExecutorTemplate.*;
 import static io.supertokens.storage.mysql.queries.EmailPasswordQueries.getQueryToCreatePasswordResetTokenExpiryIndex;
 import static io.supertokens.storage.mysql.queries.EmailPasswordQueries.getQueryToCreatePasswordResetTokensTable;
 import static io.supertokens.storage.mysql.queries.EmailVerificationQueries.*;
@@ -415,6 +415,15 @@ public class GeneralQueries {
             update(con, TOTPQueries.getQueryToCreateUsedCodesTable(start), NO_OP_SETTER);
             // index:
             update(con, TOTPQueries.getQueryToCreateUsedCodesExpiryTimeIndex(start), NO_OP_SETTER);
+        }
+
+        if (!doesTableExists(start, con, Config.getConfig(start).getBulkImportUsersTable())) {
+            getInstance(start).addState(CREATING_NEW_TABLE, null);
+            update(start, BulkImportQueries.getQueryToCreateBulkImportUsersTable(start), NO_OP_SETTER);
+            // index:
+            update(start, BulkImportQueries.getQueryToCreateStatusUpdatedAtIndex(start), NO_OP_SETTER);
+            update(start, BulkImportQueries.getQueryToCreatePaginationIndex1(start), NO_OP_SETTER);
+            update(start, BulkImportQueries.getQueryToCreatePaginationIndex2(start), NO_OP_SETTER);
         }
 
         if (!doesTableExists(start, con, Config.getConfig(start).getOAuthClientsTable())) {
@@ -987,6 +996,32 @@ public class GeneralQueries {
         }
     }
 
+    public static void makePrimaryUsers_Transaction(Start start, Connection sqlCon, AppIdentifier appIdentifier,
+                                                    List<String> userIds)
+            throws SQLException, StorageQueryException {
+
+        String users_update_QUERY = "UPDATE " + Config.getConfig(start).getUsersTable() +
+                " SET is_linked_or_is_a_primary_user = true WHERE app_id = ? AND user_id = ?";
+        String appid_to_userid_update_QUERY = "UPDATE " + Config.getConfig(start).getAppIdToUserIdTable() +
+                " SET is_linked_or_is_a_primary_user = true WHERE app_id = ? AND user_id = ?";
+
+        List<PreparedStatementValueSetter> usersUpdateBatch = new ArrayList<>();
+        List<PreparedStatementValueSetter> appIdToUserIdUpdateBatch = new ArrayList<>();
+
+        for(String userId: userIds){
+            usersUpdateBatch.add(pst -> {
+                pst.setString(1, appIdentifier.getAppId());
+                pst.setString(2, userId);
+            });
+            appIdToUserIdUpdateBatch.add(pst -> {
+                pst.setString(1, appIdentifier.getAppId());
+                pst.setString(2, userId);
+            });
+        }
+        executeBatch(sqlCon, users_update_QUERY, usersUpdateBatch);
+        executeBatch(sqlCon, appid_to_userid_update_QUERY, appIdToUserIdUpdateBatch);
+    }
+
     public static void linkAccounts_Transaction(Start start, Connection sqlCon, AppIdentifier appIdentifier,
                                                 String recipeUserId, String primaryUserId)
             throws SQLException, StorageQueryException {
@@ -1015,6 +1050,69 @@ public class GeneralQueries {
                 pst.setString(3, recipeUserId);
             });
         }
+    }
+
+    public static void linkMultipleAccounts_Transaction(Start start, Connection sqlCon, AppIdentifier appIdentifier,
+                                                        Map<String, String> recipeUserIdToPrimaryUserId)
+            throws SQLException, StorageQueryException {
+
+        if(recipeUserIdToPrimaryUserId == null || recipeUserIdToPrimaryUserId.isEmpty()){
+            return;
+        }
+
+        String update_users_QUERY = "UPDATE " + Config.getConfig(start).getUsersTable() +
+                " SET is_linked_or_is_a_primary_user = true, primary_or_recipe_user_id = ? WHERE app_id = ? AND " +
+                "user_id = ?";
+
+        String update_appid_to_userid_QUERY = "UPDATE " + Config.getConfig(start).getAppIdToUserIdTable() +
+                " SET is_linked_or_is_a_primary_user = true, primary_or_recipe_user_id = ? WHERE app_id = ? AND " +
+                "user_id = ?";
+
+        List<PreparedStatementValueSetter> usersUpdateBatch = new ArrayList<>();
+        List<PreparedStatementValueSetter> appIdToUserIdUpdateBatch = new ArrayList<>();
+
+        for(Map.Entry<String, String> linkEntry : recipeUserIdToPrimaryUserId.entrySet()) {
+            String primaryUserId = linkEntry.getValue();
+            String recipeUserId = linkEntry.getKey();
+
+            usersUpdateBatch.add(pst -> {
+                pst.setString(1, primaryUserId);
+                pst.setString(2, appIdentifier.getAppId());
+                pst.setString(3, recipeUserId);
+            });
+            appIdToUserIdUpdateBatch.add(pst -> {
+                pst.setString(1, primaryUserId);
+                pst.setString(2, appIdentifier.getAppId());
+                pst.setString(3, recipeUserId);
+            });
+        }
+
+        executeBatch(sqlCon, update_users_QUERY, usersUpdateBatch);
+        executeBatch(sqlCon, update_appid_to_userid_QUERY, appIdToUserIdUpdateBatch);
+
+        updateTimeJoinedForPrimaryUsers_Transaction(start, sqlCon, appIdentifier,
+                new ArrayList<>(recipeUserIdToPrimaryUserId.values()));
+    }
+
+    public static void updateTimeJoinedForPrimaryUsers_Transaction(Start start, Connection sqlCon,
+                                                                   AppIdentifier appIdentifier, List<String> primaryUserIds)
+            throws SQLException, StorageQueryException {
+
+        String QUERY = "UPDATE " + Config.getConfig(start).getUsersTable() +
+                " JOIN (SELECT primary_or_recipe_user_id, MIN(time_joined) AS min_time_joined FROM " +
+                Config.getConfig(start).getUsersTable() + " WHERE app_id = ? AND primary_or_recipe_user_id = ? )" +
+                " AS temp_table ON " + Config.getConfig(start).getUsersTable() + ".primary_or_recipe_user_id = " +
+                "temp_table.primary_or_recipe_user_id " +
+                " SET primary_or_recipe_user_time_joined = min_time_joined";
+
+        List<PreparedStatementValueSetter> updateBatch = new ArrayList<>();
+        for(String primaryUserId : primaryUserIds) {
+            updateBatch.add(pst -> {
+                pst.setString(1, appIdentifier.getAppId());
+                pst.setString(2, primaryUserId);
+            });
+        }
+        executeBatch(sqlCon, QUERY, updateBatch);
     }
 
     public static void unlinkAccounts_Transaction(Start start, Connection sqlCon, AppIdentifier appIdentifier,
@@ -1342,6 +1440,17 @@ public class GeneralQueries {
 
         return userIdToAuthRecipeUserInfo.keySet().stream().map(userIdToAuthRecipeUserInfo::get)
                 .collect(Collectors.toList());
+    }
+
+    public static List<AuthRecipeUserInfo> getPrimaryUserInfosForUserIds_Transaction(Start start, Connection con,
+                                                                                     AppIdentifier appIdentifier, List<String> ids)
+            throws SQLException, StorageQueryException {
+
+        List<AuthRecipeUserInfo> result = getPrimaryUserInfoForUserIds_Transaction(start, con, appIdentifier, ids);
+        if (result.isEmpty()) {
+            return null;
+        }
+        return result;
     }
 
     private static List<AuthRecipeUserInfo> getPrimaryUserInfoForUserIds_Transaction(Start start, Connection sqlCon,
@@ -1675,6 +1784,24 @@ public class GeneralQueries {
             pst.setString(1, appIdentifier.getAppId());
         }, result -> {
             return result.next();
+        });
+    }
+
+    public static List<String> findUserIdsThatExist(Start start, AppIdentifier appIdentifier, List<String> userIds)
+            throws SQLException, StorageQueryException {
+        String QUERY = "SELECT user_id FROM " + Config.getConfig(start).getAppIdToUserIdTable()
+                + " WHERE app_id = ? AND user_id IN ("+ Utils.generateCommaSeperatedQuestionMarks(userIds.size()) +")";
+        return execute(start, QUERY, pst -> {
+            pst.setString(1, appIdentifier.getAppId());
+            for(int i = 0; i<userIds.size(); i++) {
+                pst.setString(2 + i, userIds.get(i));
+            }
+        }, result -> {
+            List<String> foundUserIds = new ArrayList<>();
+            while(result.next()){
+                foundUserIds.add(result.getString(1));
+            }
+            return foundUserIds;
         });
     }