+++ src/main/java/io/supertokens/storage/mysql/queries/UserIdMappingQueries.java
@@ -20,8 +20,11 @@ import io.supertokens.pluginInterface.RowMapper;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.pluginInterface.useridmapping.UserIdMapping;
+import io.supertokens.storage.mysql.ConnectionPool;
+import io.supertokens.storage.mysql.PreparedStatementValueSetter;
 import io.supertokens.storage.mysql.Start;
 import io.supertokens.storage.mysql.config.Config;
+import io.supertokens.storage.mysql.utils.Utils;
 
 import javax.annotation.Nullable;
 import java.sql.Connection;
@@ -30,9 +33,9 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-import static io.supertokens.storage.mysql.QueryExecutorTemplate.execute;
-import static io.supertokens.storage.mysql.QueryExecutorTemplate.update;
+import static io.supertokens.storage.mysql.QueryExecutorTemplate.*;
 import static io.supertokens.storage.mysql.config.Config.getConfig;
 
 public class UserIdMappingQueries {
@@ -84,6 +87,25 @@ public class UserIdMappingQueries {
         });
     }
 
+    public static void createBulkUserIdMapping(Start start, AppIdentifier appIdentifier,
+                                               Map<String, String> superTokensUserIdToExternalUserId)
+            throws SQLException, StorageQueryException {
+        String QUERY = "INSERT INTO " + Config.getConfig(start).getUserIdMappingTable()
+                + " (app_id, supertokens_user_id, external_user_id)" + " VALUES(?, ?, ?)";
+
+        Connection sqlConnection = ConnectionPool.getConnection(start);
+        List<PreparedStatementValueSetter> setters = new ArrayList<>();
+
+        for(String superTokensUserId : superTokensUserIdToExternalUserId.keySet()) {
+            setters.add(pst -> {
+                pst.setString(1, appIdentifier.getAppId());
+                pst.setString(2, superTokensUserId);
+                pst.setString(3, superTokensUserIdToExternalUserId.get(superTokensUserId));
+            });
+        }
+        executeBatch(sqlConnection, QUERY, setters);
+    }
+
     public static UserIdMapping getUserIdMappingWithExternalUserId(Start start, AppIdentifier appIdentifier,
                                                                    String userId)
             throws SQLException, StorageQueryException {
@@ -98,6 +120,50 @@ public class UserIdMappingQueries {
                 return UserIdMappingRowMapper.getInstance().mapOrThrow(result);
             }
             return null;
+        });
+    }
+
+    public static List<UserIdMapping> getMultipleUserIdMappingWithExternalUserId_Transaction(Start start, Connection sqlCon,
+                                                                                             AppIdentifier appIdentifier,
+                                                                                             List<String> userId)
+            throws SQLException, StorageQueryException {
+        String QUERY = "SELECT * FROM " + Config.getConfig(start).getUserIdMappingTable()
+                + " WHERE app_id = ? AND external_user_id IN ( "+ Utils.generateCommaSeperatedQuestionMarks(
+                userId.size()) + " )";
+
+        return execute(sqlCon, QUERY, pst -> {
+            pst.setString(1, appIdentifier.getAppId());
+            for(int i = 0; i < userId.size(); i++) {
+                pst.setString(2 + i, userId.get(i));
+            }
+        }, result -> {
+            List<UserIdMapping> results =  new ArrayList<>();
+            while (result.next()) {
+                results.add(UserIdMappingRowMapper.getInstance().mapOrThrow(result));
+            }
+            return results;
+        });
+    }
+
+    public static List<UserIdMapping> getMultipleUserIdMappingWithSupertokensUserId_Transaction(Start start, Connection sqlCon,
+                                                                                                AppIdentifier appIdentifier,
+                                                                                                List<String> userId)
+            throws SQLException, StorageQueryException {
+        String QUERY = "SELECT * FROM " + Config.getConfig(start).getUserIdMappingTable()
+                + " WHERE app_id = ? AND supertokens_user_id IN ( "+ Utils.generateCommaSeperatedQuestionMarks(
+                userId.size()) + " )";
+
+        return execute(sqlCon, QUERY, pst -> {
+            pst.setString(1, appIdentifier.getAppId());
+            for(int i = 0; i < userId.size(); i++) {
+                pst.setString(2 + i, userId.get(i));
+            }
+        }, result -> {
+            List<UserIdMapping> results =  new ArrayList<>();
+            while (result.next()) {
+                results.add(UserIdMappingRowMapper.getInstance().mapOrThrow(result));
+            }
+            return results;
         });
     }