+++ CHANGELOG.md
@@ -7,7 +7,35 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [7.3.0]
+
+- Adds queries for Bulk Import
+- Adds support for multithreaded bulk import
 - Optimize getUserIdMappingWithEitherSuperTokensUserIdOrExternalUserId query
+
+### Migration
+
+```sql
+CREATE TABLE IF NOT EXISTS bulk_import_users (
+                id CHAR(36),
+                app_id VARCHAR(64) NOT NULL DEFAULT 'public',
+                primary_user_id VARCHAR(36),
+                raw_data TEXT NOT NULL,
+                status VARCHAR(128) DEFAULT 'NEW',
+                error_msg TEXT,
+                created_at BIGINT UNSIGNED NOT NULL, 
+                updated_at BIGINT UNSIGNED NOT NULL, 
+                PRIMARY KEY (app_id, id),
+                FOREIGN KEY(app_id) REFERENCES apps(app_id) ON DELETE CASCADE
+);
+
+CREATE INDEX bulk_import_users_status_updated_at_index ON bulk_import_users (app_id, status, updated_at);
+
+CREATE INDEX bulk_import_users_pagination_index1 ON bulk_import_users (app_id, status, created_at DESC,
+ id DESC);
+ 
+CREATE INDEX bulk_import_users_pagination_index2 ON bulk_import_users (app_id, created_at DESC, id DESC);
+```
 
 ## [7.2.0] - 2024-10-03